Redesigned player to accomodate chords
Now, notes are accumulated in state until a NoteOff message is encountered.  Then, a single note or sequence of notes are played (as appropriate) if any have accumulated since the last NoteOff.  Time delay of this NoteOff message is also recorded in state; when we get a response back from SoundFont that something has been played, we delay for this amount of time before stepping to the next event. This paces the tune.

Removed the unwanted Performance.elm.